+++ packages/flutter_spanned_controller/CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use `Range` in `AttributeSpan` and related classes to avoid confusion with UTF-16 indices used by `TextRange`.
 - Renamed `SpanList.spans` to `iter` to prevent confusing `spans.spans`.
 - `SpannedTextEditingController` takes a `SpanList` now instead of `Iterable<AttributeSpan>`.
+- Renamed `InsertBehavior` to `ExpandRule` and `FullInsertBehavior` to `SpanExpandRules`.
+  Finally I've come up with the right name for this concept :)
 
 ## [0.2.0-0] — 2020-12-14